+++ Cargo.toml
@@ -47,7 +47,6 @@ members = [
   "crates/concurrency-bounded-map-core",
   "crates/concurrency-caps-contract-core",
   "crates/concurrency-init-bootstrap-policy-core",
-  "crates/parser-backend-core",
   "tool",
   "crates/common-syntax-core",
   "common",
