+++ crates/feature-policy-core/src/lib.rs
@@ -9,7 +9,109 @@
 
 use core::fmt::{self, Display, Formatter};
 
-pub use adze_parser_backend_core::{ParserBackend, ParserBackendSelection};
+/// Message emitted when conflict handling requires GLR support.
+pub const CONFLICTS_REQUIRE_GLR_MESSAGE: &str = "Grammar has conflicts but GLR feature is not enabled. Enable the 'glr' feature in Cargo.toml or use the tree-sitter C runtime.";
+
+/// Contract outcome for backend selection under a given conflict condition.
+///
+/// This keeps behavior assertions in one place across parser and governance
+/// contracts without forcing callers to duplicate panic-string checks.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ParserBackendSelection {
+    /// A concrete backend can be selected for the current feature set.
+    Backend(ParserBackend),
+    /// Conflict grammars require the `glr` feature to be enabled.
+    ConflictsRequireGlr,
+}
+
+/// Parser backend supported by the runtime feature matrix.
+///
+/// # Examples
+///
+/// ```
+/// use adze_feature_policy_core::ParserBackend;
+///
+/// let backend = ParserBackend::GLR;
+/// assert!(backend.is_glr());
+/// assert!(backend.is_pure_rust());
+/// assert_eq!(backend.name(), "pure-Rust GLR parser");
+/// ```
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ParserBackend {
+    /// Tree-sitter C runtime (default when pure-Rust is disabled).
+    TreeSitter,
+    /// Pure Rust LR parser (simple grammars without conflicts).
+    PureRust,
+    /// Pure Rust GLR parser (conflict-capable).
+    GLR,
+}
+
+impl ParserBackend {
+    /// Resolve the backend-selection contract for a conflict condition.
+    ///
+    /// This mirrors `select` but returns an explicit outcome rather than panicking.
+    #[must_use]
+    pub const fn select_contract(has_conflicts: bool) -> ParserBackendSelection {
+        match (cfg!(feature = "glr"), cfg!(feature = "pure-rust")) {
+            (true, _) => ParserBackendSelection::Backend(Self::GLR),
+            (false, true) => {
+                if has_conflicts {
+                    ParserBackendSelection::ConflictsRequireGlr
+                } else {
+                    ParserBackendSelection::Backend(Self::PureRust)
+                }
+            }
+            _ => ParserBackendSelection::Backend(Self::TreeSitter),
+        }
+    }
+
+    /// Select parser backend based on feature flags and grammar metadata.
+    ///
+    /// # Arguments
+    ///
+    /// * `has_conflicts` - Whether the grammar contains shift/reduce or reduce/reduce conflicts.
+    ///
+    /// # Panics
+    ///
+    /// Panics if `has_conflicts` is true and the `pure-rust` feature is enabled without the `glr` feature.
+    #[must_use]
+    pub const fn select(has_conflicts: bool) -> Self {
+        match Self::select_contract(has_conflicts) {
+            ParserBackendSelection::Backend(backend) => backend,
+            ParserBackendSelection::ConflictsRequireGlr => {
+                panic!("{}", CONFLICTS_REQUIRE_GLR_MESSAGE)
+            }
+        }
+    }
+
+    /// Whether this backend is the GLR parser.
+    #[must_use]
+    pub const fn is_glr(self) -> bool {
+        matches!(self, Self::GLR)
+    }
+
+    /// Whether this backend is a pure-Rust parser (LR or GLR).
+    #[must_use]
+    pub const fn is_pure_rust(self) -> bool {
+        matches!(self, Self::PureRust | Self::GLR)
+    }
+
+    /// Human-readable backend name.
+    #[must_use]
+    pub const fn name(self) -> &'static str {
+        match self {
+            Self::TreeSitter => "tree-sitter C runtime",
+            Self::PureRust => "pure-Rust LR parser",
+            Self::GLR => "pure-Rust GLR parser",
+        }
+    }
+}
+
+impl Display for ParserBackend {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.name())
+    }
+}
 
 /// Snapshot of parser-related feature flags for this build.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
@@ -145,6 +247,48 @@ mod tests {
         );
         assert_eq!(ParserBackend::PureRust.to_string(), "pure-Rust LR parser");
         assert_eq!(ParserBackend::GLR.to_string(), "pure-Rust GLR parser");
+    }
+
+    #[test]
+    fn backend_name_returns_human_readable_string() {
+        assert_eq!(ParserBackend::TreeSitter.name(), "tree-sitter C runtime");
+        assert_eq!(ParserBackend::PureRust.name(), "pure-Rust LR parser");
+        assert_eq!(ParserBackend::GLR.name(), "pure-Rust GLR parser");
+    }
+
+    #[test]
+    fn backend_is_glr_only_for_glr() {
+        assert!(ParserBackend::GLR.is_glr());
+        assert!(!ParserBackend::PureRust.is_glr());
+        assert!(!ParserBackend::TreeSitter.is_glr());
+    }
+
+    #[test]
+    fn backend_is_pure_rust_for_lr_and_glr() {
+        assert!(ParserBackend::PureRust.is_pure_rust());
+        assert!(ParserBackend::GLR.is_pure_rust());
+        assert!(!ParserBackend::TreeSitter.is_pure_rust());
+    }
+
+    #[test]
+    fn backend_select_matches_feature_contract() {
+        #[cfg(feature = "glr")]
+        {
+            assert_eq!(ParserBackend::select(false), ParserBackend::GLR);
+            assert_eq!(ParserBackend::select(true), ParserBackend::GLR);
+        }
+
+        #[cfg(all(feature = "pure-rust", not(feature = "glr")))]
+        {
+            assert_eq!(ParserBackend::select(false), ParserBackend::PureRust);
+            assert!(std::panic::catch_unwind(|| ParserBackend::select(true)).is_err());
+        }
+
+        #[cfg(not(any(feature = "pure-rust", feature = "glr")))]
+        {
+            assert_eq!(ParserBackend::select(false), ParserBackend::TreeSitter);
+            assert_eq!(ParserBackend::select(true), ParserBackend::TreeSitter);
+        }
     }
 
     #[test]
