+++ crates/feature-policy-core/tests/integration_parser_backend_bridge.rs
@@ -1,7 +1,6 @@
 use std::panic::catch_unwind;
 
-use adze_feature_policy_core::ParserFeatureProfile;
-use adze_parser_backend_core::ParserBackend as Backend;
+use adze_feature_policy_core::{ParserBackend, ParserFeatureProfile};
 
 #[test]
 fn parser_backend_reexport_and_profile_resolve_backend_stay_in_sync() {
@@ -9,17 +8,17 @@ fn parser_backend_reexport_and_profile_resolve_backend_stay_in_sync() {
 
     for has_conflicts in [false, true] {
         let from_profile = catch_unwind(|| profile.resolve_backend(has_conflicts));
-        let from_reexport = catch_unwind(|| Backend::select(has_conflicts));
+        let from_backend = catch_unwind(|| ParserBackend::select(has_conflicts));
 
         assert_eq!(
             from_profile.is_ok(),
-            from_reexport.is_ok(),
+            from_backend.is_ok(),
             "panic behavior differs for has_conflicts={has_conflicts}"
         );
 
-        if let (Ok(from_profile), Ok(from_reexport)) = (from_profile, from_reexport) {
+        if let (Ok(from_profile), Ok(from_backend)) = (from_profile, from_backend) {
             assert_eq!(
-                from_profile, from_reexport,
+                from_profile, from_backend,
                 "backend selection changed for has_conflicts={has_conflicts}"
             );
         }
