--- crates/parser-backend-core/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "adze-parser-backend-core"
-version = "0.1.0"
-edition.workspace = true
-rust-version.workspace = true
-license = "MIT OR Apache-2.0"
-description = "Pure parser backend selection primitives."
-repository = "https://github.com/EffortlessMetrics/adze"
-publish = ["crates.io"]
-keywords = ["parser", "grammar", "tree-sitter", "ast", "glr"]
-categories = ["parsing", "development-tools"]
-homepage = "https://github.com/EffortlessMetrics/adze"
-documentation = "https://docs.rs/adze-parser-backend-core"
-
-[features]
-default = []
-pure-rust = []
-tree-sitter-c2rust = []
-tree-sitter-standard = []
-glr = ["pure-rust"]
-strict_api = []
-strict_docs = []
-
-[dev-dependencies]
-criterion.workspace = true
-proptest.workspace = true
-
-[[bench]]
-name = "parser_backend_bench"
-harness = false
