--- crates/parser-backend-core/README.md
@@ -1,10 +1,0 @@
-# adze-parser-backend-core
-
-Pure parser backend selection primitives.
-
-Part of the [adze](https://github.com/EffortlessMetrics/adze) workspace.
-
-## License
-
-Licensed under either of [Apache License, Version 2.0](../../LICENSE-APACHE)
-or [MIT License](../../LICENSE-MIT) at your option.
