+++ policy/package-boundary.toml
@@ -657,20 +657,6 @@ last_changed = "2026-05-12"
 last_changed_pr = "#694"
 
 [[package]]
-name = "adze-parser-backend-core"
-path = "crates/parser-backend-core"
-category = "owner-module-migration-target"
-owner = "parser-contracts"
-production_use = true
-publish_intent = "retire-or-inline"
-support_tier_impact = "temporary internal surface; no stable product claim until migrated or promoted"
-ci_impact = "collapse target; keep out of default CI expansion unless routed by owner"
-migration_target = "parser-contracts"
-removal_or_promotion_condition = "Move into the SRP owner submodule or reclassify with an accepted spec/ADR before 0.9 release."
-last_changed = "2026-05-12"
-last_changed_pr = "#694"
-
-[[package]]
 name = "adze-parser-contract"
 path = "crates/parser-contract"
 category = "owner-module-migration-target"
